+++ src/components/campaign/campaignDashboard/npcsTab/NpcsTabMain.jsx
@@ -1,19 +1,18 @@
+import {Add as AddIcon, Link as LinkIcon} from '@mui/icons-material';
 import React, { useEffect } from "react";
-import { Grid, Paper } from "@mui/material";
+import {Button} from '@mui/material';
 import {useParams} from 'react-router-dom';
-import EmptyEntityList from '../../common/EmptyEntityList.jsx';
-import EntityListError from '../../common/EntityListError.jsx';
-import EntityListLoading from '../../common/EntityListLoading.jsx';
+import BrowserMain from '../../common/entityBrowser/BrowserMain.jsx';
+import EntityListError from '../../common/entityBrowser/EntityListError.jsx';
 import {useFoldersStore} from '../../stores/folderStore.js';
-import NpcsTabHeader from "./NpcsTabHeader";
-import SearchbarFilter from "./SearchbarFilter";
+import NPCsSearchFilters from "./NPCsSearchFilters.jsx";
 import LinkNpcDialog from "./LinkNpcDialog";
 import SimpleNpcDialogEdit from "./SimpleNpcDialogEdit";
-import FeedbackSnackbar from "./FeedbackSnackbar";
 import useCampaignNpcs from "./hooks/useCampaignNpcs";
 import { useNpcStore } from "./stores/npcDataStore";
 import {useNpcDialogsStore} from "./stores/npcDialogsStore";
 import NpcExplorer from "./NpcExplorer";
+import {useNpcFiltersStore} from './stores/npcFiltersStore.js';
 
 const NpcsTabMain = () => {
   const {campaignId} = useParams();
@@ -44,6 +43,17 @@ const NpcsTabMain = () => {
   } = useFoldersStore();
 
   const {
+    filterSearchText,
+    setNpcSortOrder,
+    setNpcAttitudeFilter,
+    setShowVillainsOnly,
+    setNpcRank,
+    setNpcSpecies,
+    setFilterSearchText,
+    setNpcSortDirection,
+  } = useNpcFiltersStore();
+
+  const {
     // Dialog states
     isLinkNpcDialogOpen,
     linkNpcSearchText,
@@ -71,76 +81,86 @@ const NpcsTabMain = () => {
     initializeNpcs
   ]);
 
+  const clearAllFilters = () => {
+    setFilterSearchText("");
+    setNpcSortOrder("name");
+    setNpcSortDirection("asc");
+    setNpcAttitudeFilter("all");
+    setShowVillainsOnly(false);
+    setNpcRank("");
+    setNpcSpecies("");
+  };
+
   return (
-    <Paper elevation={3} sx={{ p: 3 }}>
-      <Grid container spacing={1}>
-        {/* Header */}
-        <Grid item xs={12}>
-          <NpcsTabHeader onCreateSimpleNpc={handleOpenSimpleNpcDialogEdit} onLinkNpc={handleAddExistingNpc} />
-        </Grid>
-
-        {/* Search, sort, and filter controls */}
-        {campaignNpcs.length > 0 && (
-          <Grid item xs={12}>
-            <SearchbarFilter />
-          </Grid>
-        )}
-
-        {/* Loading state */}
-        {isLoading && (
-          <Grid item xs={12}>
-            <EntityListLoading type={'NPC'} />
-          </Grid>
-        )}
-
-        {/* Error state */}
-        {loadError && (
-          <Grid item xs={12}>
-            <EntityListError loadError={loadError} retry={loadNpcs} />
-          </Grid>
-        )}
-
-        {/* Empty state */}
-        {!isLoading && !loadError && campaignNpcs.length === 0 && (
-          <Grid item xs={12}>
-            <EmptyEntityList handleAdd={handleAddExistingNpc} type={'NPC'}></EmptyEntityList>
-          </Grid>
-        )}
-
-        {/* Display NPCs or Empty State for the current filter */}
-        {!isLoading && !loadError && campaignNpcs.length > 0 && (
+    <>
+      <BrowserMain
+        headerTitle="Campaign NPCs"
+        headerActions={
           <>
-            <NpcExplorer
-              campaignNpcs={campaignNpcs}
-              handleToggleNpc={handleToggleNpc}
-            />
+            <Button
+              variant="contained"
+              color="primary"
+              startIcon={<AddIcon />}
+              onClick={handleOpenSimpleNpcDialogEdit}
+              sx={{ mr: 1 }}
+            >
+              Add Simplified NPC
+            </Button>
+
+            <Button
+              variant="contained"
+              color="primary"
+              startIcon={<LinkIcon />}
+              onClick={handleAddExistingNpc}
+              sx={{ mr: 1 }}
+            >
+              Link NPC
+            </Button>
           </>
-        )}
+        }
+        emptyListAction={handleAddExistingNpc}
+        listErrorRetry={loadNpcs}
+        loadError={loadError}
+        isLoading={isLoading}
+        items={campaignNpcs}
+        snackbar={snackbar}
+        handleSnackbarClose={handleSnackbarClose}
+        listErrorComponent={
+          <EntityListError
+            onRetry={loadNpcs}
+            errorMessage={loadError}
+          />
+        }
+        explorerComponent={
+          <NpcExplorer
+            campaignNpcs={campaignNpcs}
+            handleToggleNpc={handleToggleNpc}
+          />
+        }
+        searchText={filterSearchText}
+        handleSearchTextUpdate={setFilterSearchText}
+        handleFiltersClearAll={clearAllFilters}
+        searchFiltersComponent={<NPCsSearchFilters />}
+      ></BrowserMain>
 
-        {/* Link NPC Dialog */}
-       <LinkNpcDialog
-          open={isLinkNpcDialogOpen}
-          handleClose={handleCloseLinkDialog}
-          searchText={linkNpcSearchText}
-          setSearchText={setLinkNpcSearchText}
-          filteredNpcs={filteredNpcsForDialog}
-          associatedNpcIds={associatedNpcIds}
-          handleToggleNpc={handleToggleNpc}
-        />
-        <SimpleNpcDialogEdit
-          open={isSimpleNpcDialogEditOpen}
-          onClose={handleCloseSimpleNpcDialogEdit}
-          onSubmit={handleCreateSimpleNpc}
-          initialNpc={simpleNpcItem}
-        />
 
-        {/* Feedback snackbar */}
-        <FeedbackSnackbar
-          snackbar={snackbar}
-          handleSnackbarClose={handleSnackbarClose}
-        />
-      </Grid>
-    </Paper>
+      {/* Link NPC Dialog */}
+      <LinkNpcDialog
+        open={isLinkNpcDialogOpen}
+        handleClose={handleCloseLinkDialog}
+        searchText={linkNpcSearchText}
+        setSearchText={setLinkNpcSearchText}
+        filteredNpcs={filteredNpcsForDialog}
+        associatedNpcIds={associatedNpcIds}
+        handleToggleNpc={handleToggleNpc}
+      />
+      <SimpleNpcDialogEdit
+        open={isSimpleNpcDialogEditOpen}
+        onClose={handleCloseSimpleNpcDialogEdit}
+        onSubmit={handleCreateSimpleNpc}
+        initialNpc={simpleNpcItem}
+      />
+    </>
   );
 };
